requirements.txt
aiogram>=3.17.0
pydantic>=2.4.1,<2.11
aiohttp>=3.9.0,<3.12
python-dotenv>=1.0.1
openai>=1.54.0
aiosqlite>=0.20.0
httpx>=0.27.0
aiofiles>=24.1.0
# kwork - берём с GitHub (зависимости будут установлены автоматически)
